fixing errors

+++ inc/util.h
@@ -18,7 +18,7 @@
 #ifndef CUTS_UTIL_H
 #define CUTS_UTIL_H
 #include "./common.h"
-unsigned long long int file_reader(std::string input_file, vector<set<unsigned int>> &ns);
+unsigned int file_reader(std::string input_file, vector<set<unsigned int>> &ns);
 
 void write_kcore_to_disk(unsigned int*, unsigned long long int, std::string file);
 #endif //CUTS_UTIL_H
